requirements.txt
langchain
python-dotenv
ipykernel
langchain-groq
langchain-google-genai
langchain_community
pypdf
faiss-cpu
structlog
PyMuPDF
langchain-openai
pandas

-e . # For installing local packages
